+++ filters/url.js
@@ -1,13 +1,17 @@
 import { join, normalize } from "../deps/path.js";
 
 export default function (site) {
-  return function (url, absolute) {
+  return function (path, absolute) {
     try {
-      return new URL(url).toString();
+      return new URL(path).toString();
     } catch (err) {
-      url = normalize(join(site.options.pathPrefix, url));
+      if (!site.options.location) {
+        return normalize(join("/", path));
+      }
 
-      return absolute ? site.options.url + url : url;
+      path = normalize(join(site.options.location.pathname, path));
+
+      return absolute ? site.options.location.origin + path : path;
     }
   };
 }