+++ loaders/module.js
@@ -1,4 +1,5 @@
 export default async function (path) {
+  console.log(`file://${path}`);
   const fileData = await import(`file://${path}`);
   const data = {};