+++ plugins/bundler.js
@@ -11,10 +11,10 @@ export default function () {
     const from = page.src.path + page.src.ext;
 
     const [diagnostics, emit] = await Deno.bundle(from, {
-      [from]: page.rendered,
+      [from]: page.content,
     });
 
-    page.rendered = emit;
+    page.content = emit;
     page.dest.ext = ".js";
   }
 }