+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.4.1] - Unreleased
+## [0.5.0] - 2020-09-24
+### Added
+- Ability to generate multiple pages using generators
+
+### Changed
+- Replaced `pathPrefix` and `url` with `location`
+
 ### Fixed
 - `url` filter bugs
 
@@ -72,7 +78,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## 0.1.0 - 2020-09-13
 First version
 
-[0.4.1]: https://github.com/oscarotero/lume/compare/v0.4.0...HEAD
+[0.5.0]: https://github.com/oscarotero/lume/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/oscarotero/lume/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/oscarotero/lume/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/oscarotero/lume/compare/v0.2.3...v0.3.0